+++ __tests__/happy.localeOnStartConversation.js
@@ -6,7 +6,8 @@ import { timeouts } from './constants.json';
 import * as createDirectLine from './setup/createDirectLine';
 import waitForBotToRespond from './setup/waitForBotToRespond';
 
-describe('Happy path', () => {
+// Skipping because the bot at dljstestbot.azurewebsites.net is not available.
+describe.skip('Happy path', () => {
   let unsubscribes;
 
   beforeEach(() => unsubscribes = []);