+++ __tests__/unhappy.postActivityFatalAfterConnect.js
@@ -10,7 +10,8 @@ import postActivity from './setup/postActivity';
 import waitForBotToEcho from './setup/waitForBotToEcho';
 import waitForConnected from './setup/waitForConnected';
 
-describe('Unhappy path', () => {
+// Skipping because the bot at dljstestbot.azurewebsites.net is not available.
+describe.skip('Unhappy path', () => {
   let unsubscribes;
 
   beforeEach(() => (unsubscribes = []));